refactor: 스토리북 추가 및 경로 변경
